# openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Generated API
  version: "1.0"
paths:
...
  /workflow/abort:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbortERezept'
      responses:
        "200":
          description: OK
  /workflow/batch-sign:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: string
      responses:
        "200":
          description: OK
...
  /workflow/comfortsignature/user-id:
    get:
      responses:
        "200":
          description: OK
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: string
      responses:
        "200":
          description: OK
...
  /workflow/sign:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: string
      responses:
        "200":
          description: OK
  /workflow/task:
    post:
      parameters:
      - name: flowtype
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
  /workflow/update:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateERezept'
      responses:
        "200":
          description: OK
...
components:
  schemas:
    AbortERezept:
      type: object
      properties:
        accessCode:
          type: string
        taskId:
          type: string
...
    UpdateERezept:
      type: object
      properties:
        accessCode:
          type: string
        signedBytes:
          type: string
        taskId:
          type: string
...